A bunch of fixes, and adding email functionality
The script wasn't actually updating filters when something had changed... it'd insert the first time, then tell you it was updating but wasn't.  Fixed some issues in the comparison logic as well as the netconf.

Added some email functionality.  You can now choose to have an email when a filter is updated AND/OR have an email when there is an error.  Particularly useful if you're running this in cron.